36_django_testing/pyproject.toml
[tool.poetry]
name = "some_shop"
version = "0.1"
description = "first project"
authors = ["Me"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.2"
celery = "^5.3.1"
django-celery-results = "^2.5.1"
django-mail-templated = "^2.6.5"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
